begun typescript everywhere
