+++ package.json
@@ -7,16 +7,21 @@
   },
   "version": "1.0.1",
   "license": "MIT",
+  "repository": {
+    "directory": "https://github.com/kwiksilver3441/cocbot"
+  },
   "scripts": {
-    "dev": "nodemon ./src/index.js",
-    "start": "yarn node ./src/deploy-commands.js; yarn node ./src/index.js",
-    "deploy": "yarn node ./src/deploy-commands.js"
+    "deploy": "yarn node ./src/deploy-commands.js",
+    "build": "tsc --build",
+    "watch": "tsc --watch",
+    "dev": "nodemon ./dist/index.js",
+    "start": "yarn node ./src/deploy-commands.js; yarn node ./src/index.js"
   },
   "dependencies": {
     "@discordjs/builders": "^0.11.0",
     "@discordjs/rest": "^0.2.0-canary.0",
     "@keyv/sqlite": "^3.0.0",
-    "discord-api-types": "^0.26.0",
+    "discord-api-types": "^0.27.0",
     "discord.js": "^13.6.0",
     "dotenv": "^14.1.0",
     "i18n": "^0.13.4",
@@ -25,11 +30,17 @@
   },
   "devDependencies": {
     "@types/i18n": "^0.13.2",
+    "@types/keyv": "^3.1.3",
+    "@types/lodash": "^4.14.178",
+    "@types/node": "^17.0.17",
+    "@typescript-eslint/eslint-plugin": "^5.11.0",
+    "@typescript-eslint/parser": "^5.11.0",
     "eslint": "^7.32.0",
     "eslint-config-standard": "^16.0.3",
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^5.2.0",
-    "nodemon": "^2.0.15"
+    "nodemon": "^2.0.15",
+    "typescript": "^4.5.5"
   }
 }
