+++ src/bot.ts
@@ -1,8 +1,8 @@
-const { Client, Intents, Collection } = require('discord.js')
-const { readdirSync } = require('fs')
-const { join } = require('path')
-const { getLang } = require('./util/')
-const { i18n } = require('./plugins/')
+import { Client, Intents, Collection } from 'discord.js'
+import { readdirSync } from 'fs'
+import { join } from 'path'
+import { getLang } from './util/'
+import { i18n } from './plugins/'
 
 const client = new Client({ intents: [Intents.FLAGS.GUILDS] })
 
@@ -15,7 +15,7 @@ for (const file of commandFiles) {
 }
 
 client.once('ready', () => {
-	client.user.setActivity('/ slash commands', { type: 'WATCHING' })
+	client.user!.setActivity('/ slash commands', { type: 'WATCHING' })
 	console.log('Ready')
 })
 
@@ -26,7 +26,7 @@ client.on('interactionCreate', async (interaction) => {
 	try {
 		await command.execute(interaction)
 	}
-	catch (err) {
+	catch (err: any) {
 		console.error({ error: err, interaction })
 		return await interaction.reply({ content: 'There was an error while trying to execute this command\nError was logged', ephemeral: true })
 	}
@@ -39,7 +39,7 @@ client.on('interactionCreate', async (interaction) => {
 	try {
 		await command.buttonExecute(interaction)
 	}
-	catch (err) {
+	catch (err: any) {
 		console.error({ error: err, interaction })
 		return await interaction.reply({ content: 'There was an error while trying to execute this command\nError was logged', ephemeral: true })
 	}
