+++ src/commands/add.ts
@@ -1,9 +1,10 @@
-const { SlashCommandBuilder } = require('@discordjs/builders')
-const { unlink } = require('fs/promises')
-const { addModifiers, buildTempFile } = require('../util/')
-const { MessageAttachment } = require('discord.js')
+import { SlashCommandBuilder } from '@discordjs/builders'
+import { unlink } from 'fs/promises'
+import { addModifiers, buildTempFile } from '../util/'
+import { CommandInteraction, MessageAttachment } from 'discord.js'
+import { i18n } from '../plugins/'
 
-module.exports = {
+export default {
 	data: new SlashCommandBuilder()
 		.setName('add')
 		.setDescription('Adds multiple inputs')
@@ -17,12 +18,11 @@ module.exports = {
 		.addBooleanOption((option) =>
 			option.setName('ephemeral')
 				.setDescription('Hides the value for only you to see')),
-	async execute(interaction) {
-		const { i18n } = require('../plugins/')
+	async execute(interaction: CommandInteraction) {
 		let gFile
 		const ephemeral = interaction.options.getBoolean('ephemeral') ?? false
 		const modifiers = interaction.options.getString('modifiers') ?? ''
-		const values = interaction.options.getString('values')
+		const values = interaction.options.getString('values')!
 		const nums = values.split(/\s+/).filter((element) => !isNaN(element))
 		if (values === '') {
 			return await interaction.reply({ content: i18n.__('valueNotEmpty'), ephemeral: true })
@@ -30,7 +30,7 @@ module.exports = {
 		else if (!nums.length) {
 			return await interaction.reply({ content: i18n.__('valueYesNumbers'), ephemeral: true })
 		}
-		else if (nums.length > Number.parseInt(process.env.MAX_SAFE_ARGS)) {
+		else if (nums.length > Number.parseInt(process.env.MAX_SAFE_ARGS!)) {
 			console.warn(interaction)
 			return await interaction.reply({ content: i18n.__('valueOverMaxSafeArgs'), ephemeral: true })
 		}
@@ -57,14 +57,14 @@ module.exports = {
 				const mFile = new MessageAttachment(file)
 				return await interaction.reply({ content: i18n.__('totalIs', { value: obj.total.toLocaleString(interaction.locale) }), ephemeral, files: [mFile] })
 			}
-			catch (err) {
+			catch (err: any) {
 				console.error({ error: err, interaction })
 				return await interaction.reply({ content: `${i18n.__('error')}: ${err.message}`, ephemeral: true })
 			}
 			finally {
 				if (gFile != null) {
 					unlink(gFile)
-						.catch((err) => {
+						.catch((err: any) => {
 							console.error({ error: err, interaction })
 						})
 				}
