+++ src/commands/addToStat.ts
@@ -1,7 +1,9 @@
-const { SlashCommandBuilder } = require('@discordjs/builders')
-const { addToStat } = require('../util/')
+import { SlashCommandBuilder } from '@discordjs/builders'
+import { CommandInteraction } from 'discord.js'
+import { addToStat } from '../util/'
+import { i18n } from '../plugins/'
 
-module.exports = {
+export default {
 	data: new SlashCommandBuilder()
 		.setName('add-to-stat')
 		.setDescription('Returns how many hops you need to add a stat to a specified amount')
@@ -16,10 +18,9 @@ module.exports = {
 		.addIntegerOption((option) =>
 			option.setName('ephemeral')
 				.setDescription('Hides the value for only you to see')),
-	async execute(interaction) {
-		const { i18n } = require('../plugins/')
-		const from = interaction.options.getInteger('from')
-		const to = interaction.options.getInteger('to')
+	async execute(interaction: CommandInteraction) {
+		const from = interaction.options.getInteger('from')!
+		const to = interaction.options.getInteger('to')!
 		const ephemeral = interaction.options.getBoolean('ephemeral') ?? false
 		if (from < 0) {
 			return await interaction.reply({ content: i18n.__('fromNotNegative'), ephemeral: true })
@@ -35,7 +36,7 @@ module.exports = {
 				content = `${content}\n${i18n.__('fromLevel')} ${from.toLocaleString(interaction.locale)}`
 				return await interaction.reply({ content, ephemeral })
 			}
-			catch (err) {
+			catch (err: any) {
 				console.error({ error: err, interaction })
 				return await interaction.reply({ content: `${i18n.__('error')}: ${err.message}`, ephemeral: true })
 			}
