+++ src/commands/multiply.ts
@@ -1,15 +1,16 @@
-const { SlashCommandBuilder } = require('@discordjs/builders')
-const { unlink } = require('fs/promises')
-const { addModifiers, buildTempFile } = require('../util/')
-const { MessageAttachment } = require('discord.js')
+import { SlashCommandBuilder } from '@discordjs/builders'
+import { unlink } from 'fs/promises'
+import { addModifiers, buildTempFile } from '../util/'
+import { CommandInteraction, MessageAttachment } from 'discord.js'
+import { i18n } from '../plugins/'
 
-module.exports = {
+export default {
 	data: new SlashCommandBuilder()
-		.setName('subtract')
-		.setDescription('Subtracts multiple inputs')
+		.setName('multiply')
+		.setDescription('Multiplies multiple inputs')
 		.addStringOption((option) =>
 			option.setName('values')
-				.setDescription('Each value you wish to subtract separated by a space, ex: 10 20 30 40')
+				.setDescription('Each value you wish to multiply separated by a space, ex: 10 20 30 40')
 				.setRequired(true))
 		.addStringOption((option) =>
 			option.setName('modifiers')
@@ -17,12 +18,11 @@ module.exports = {
 		.addBooleanOption((option) =>
 			option.setName('ephemeral')
 				.setDescription('Hides the value for only you to see')),
-	async execute(interaction) {
-		const { i18n } = require('../plugins/')
+	async execute(interaction: CommandInteraction) {
 		let gFile
 		const ephemeral = interaction.options.getBoolean('ephemeral') ?? false
 		const modifiers = interaction.options.getString('modifiers') ?? ''
-		const values = interaction.options.getString('values')
+		const values = interaction.options.getString('values')!
 		const nums = values.split(/\s+/).filter((element) => !isNaN(element))
 		if (values === '') {
 			return await interaction.reply({ content: i18n.__('valueNotEmpty'), ephemeral: true })
@@ -44,7 +44,7 @@ module.exports = {
 				nums.forEach((element, index) => {
 					if (index !== 0) {
 						element = Number.parseInt(element)
-						obj.value = obj.value - element
+						obj.value = obj.value * element
 						obj.total = obj.value
 					}
 				})
@@ -59,7 +59,7 @@ module.exports = {
 				const mFile = new MessageAttachment(file)
 				return await interaction.reply({ content: i18n.__('totalIs', { value: obj.total.toLocaleString(interaction.locale) }), ephemeral, files: [mFile] })
 			}
-			catch (err) {
+			catch (err: any) {
 				console.error({ error: err, interaction })
 				return await interaction.reply({ content: `${i18n.__('error')}: ${err.message}`, ephemeral: true })
 			}
