+++ src/commands/roll.ts
@@ -1,16 +1,20 @@
-const { SlashCommandBuilder } = require('@discordjs/builders')
-const { buildTempFile, roll, rollsWriteContent } = require('../util/')
-const { unlink } = require('fs/promises')
-const { MessageAttachment } = require('discord.js')
+import { SlashCommandBuilder } from '@discordjs/builders'
+import { buildTempFile, roll, rollsWriteContent } from '../util/'
+import { unlink } from 'fs/promises'
+import { CommandInteraction, MessageAttachment } from 'discord.js'
+import { i18n } from '../plugins/'
 
-
-module.exports = {
+export default {
 	data: new SlashCommandBuilder()
-		.setName('roll-one')
-		.setDescription('Rolls one die')
+		.setName('roll')
+		.setDescription('Rolls dice')
 		.addIntegerOption((option) =>
 			option.setName('size')
-				.setDescription('Size of the die')
+				.setDescription('Size of each die')
+				.setRequired(true))
+		.addIntegerOption((option) =>
+			option.setName('number')
+				.setDescription('How many of each die')
 				.setRequired(true))
 		.addStringOption((option) =>
 			option.setName('modifiers')
@@ -23,21 +27,24 @@ module.exports = {
 				.setDescription('Causes a reroll when the roll value hits or exceeds target'))
 		.addIntegerOption((option) =>
 			option.setName('rerolls')
-				.setDescription('Rerolls the die x amount of times'))
+				.setDescription('Rerolls the same size and number of dice x amount of times'))
 		.addBooleanOption((option) =>
 			option.setName('ephemeral')
 				.setDescription('Hides the value for only you to see')),
-	async execute(interaction) {
-		const { i18n } = require('../plugins/')
+	async execute(interaction: CommandInteraction) {
 		let gFile
-		const size = interaction.options.getInteger('size')
-		const rerolls = interaction.options.getInteger('rerolls') ?? 1
+		const number = interaction.options.getInteger('number')!
+		const size = interaction.options.getInteger('size')!
 		const ephemeral = interaction.options.getBoolean('ephemeral') ?? false
 		const explode = interaction.options.getInteger('explode') ?? size + 1
 		const modifiers = interaction.options.getString('modifiers') ?? ''
 		const diceModifiers = interaction.options.getString('dice-modifiers') ?? ''
+		const rerolls = interaction.options.getInteger('rerolls') ?? 1
 		if (size < 1) {
 			return await interaction.reply({ content: i18n.__('sizeNegativeOrZero'), ephemeral: true })
+		}
+		else if (number < 1) {
+			return await interaction.reply({ content: i18n.__('numberNegativeOrZero'), ephemeral: true })
 		}
 		else if (explode > size + 1) {
 			return await interaction.reply({ content: i18n.__('explodeOverValue'), ephemeral: true })
@@ -50,14 +57,14 @@ module.exports = {
 		}
 		else {
 			try {
-				const obj = await roll({ size, number: 1, rerolls, explode, diceModifiers })
+				const obj = await roll({ size, number, rerolls, explode, diceModifiers })
 				const content = await rollsWriteContent(obj, modifiers)
 				const file = await buildTempFile(JSON.stringify(content, null, 2))
 				gFile = file
 				const mFile = new MessageAttachment(file)
 				return await interaction.reply({ content: i18n.__('totalIs', { value: content.total.toLocaleString(interaction.locale) }), ephemeral, files: [mFile] })
 			}
-			catch (err) {
+			catch (err: any) {
 				console.error({ error: err, interaction })
 				return await interaction.reply({ content: `${i18n.__('error')}: ${err.message}`, ephemeral: true })
 			}
