--- src/plugins/i18n.js
@@ -1,9 +1,0 @@
-const { I18n } = require('i18n')
-
-const i18n = new I18n({
-	locales: ['en', 'es'],
-	directory: './lang',
-	retryInDefaultLocale: true
-})
-
-module.exports = i18n
