+++ src/plugins/i18n.ts
@@ -1,0 +1,7 @@
+import { I18n } from 'i18n'
+
+export default new I18n(({
+	locales: ['en', 'es'],
+	directory: './lang',
+	retryInDefaultLocale: true
+}))
