--- src/plugins/index.js
@@ -1,6 +1,0 @@
-const keyv = require('./keyv')
-const i18n = require('./i18n')
-
-module.exports = {
-	keyv, i18n
-}
