+++ src/plugins/index.ts
@@ -1,0 +1,4 @@
+import keyv from './keyv'
+import i18n from './i18n'
+
+export { keyv, i18n }
