+++ src/types/discord.d.ts
@@ -1,0 +1,14 @@
+import { Collection } from 'discord.js'
+
+declare module 'discord.js' {
+    export interface Client {
+        commands: Collection<unknown, any>
+    }
+}
+import { Collection } from 'discord.js'
+
+declare module 'discord.js' {
+    export interface Client {
+        commands: Collection<unknown, any>
+    }
+}
