+++ src/util/addModifiers.ts
@@ -4,12 +4,12 @@
  * @param {number} value
  * @returns {number} number
  */
-module.exports = (modifiers, value) => {
+export default (modifiers: string, value: number): number => {
 	let modified = 0
 	const regex = /[+|\-|*|x|/]\d+/g
 	const match = modifiers.match(regex)
 	if (match == null) throw new Error('Incorrect values in modifiers')
-	if (match.length > process.env.MAX_SAFE_ARGS) {
+	if (match.length > Number.parseInt(process.env.MAX_SAFE_ARGS!)) {
 		throw new Error('Value is over the maximum safe arguments limit. Why are you using this many arguments?')
 	}
 	match.forEach((string) => {
