--- src/util/index.js
@@ -1,17 +1,0 @@
-const addModifiers = require('./addModifiers')
-const addToHop = require('./addToHop')
-const addToStat = require('./addToStat')
-const buildTempFile = require('./buildTempFile')
-const getLang = require('./getLang')
-const roll = require('./roll')
-const rollsWriteContent = require('./rollsWriteContent')
-
-module.exports = {
-	addModifiers,
-	addToHop,
-	addToStat,
-	buildTempFile,
-	getLang,
-	roll,
-	rollsWriteContent
-}
